TravelBlog/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-appconf==1.0.5
django-avatar==8.0.0
django-taggit==4.0.0
dnspython==2.4.2
geographiclib==2.0
geopy==2.4.0
Pillow==10.1.0
sqlparse==0.4.4
